+++ nemo/collections/nlp/modules/common/megatron/retrieval_services/util.py
@@ -1,0 +1,30 @@
+# Copyright (c) 2023, NVIDIA CORPORATION.  All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import requests
+import json
+import threading
+
+
+headers = {"Content-Type": "application/json"}
+
+lock = threading.Lock()
+
+__all__ = ["request_data", "lock"]
+
+
+def request_data(data, ip='localhost', port=None):
+    resp = requests.put(f'http://{ip}:{port}/knn', data=json.dumps(data), headers=headers)
+    return resp.json()
+
