+++ scripts/nlp_language_modeling/service_launch_scripts/start_dynamic_retrieval_service.py
@@ -14,7 +14,7 @@
 """
 """
 
-from nemo.collections.nlp.modules.common.megatron.retrieval_service import DynamicRetrievalServer
+from nemo.collections.nlp.modules.common.megatron.retrieval_services.dynamic_retrieve_server import DynamicRetrievalServer
 from nemo.collections.nlp.modules.common.tokenizer_utils import get_nmt_tokenizer
 from nemo.core.config import hydra_runner
