+++ scripts/nlp_language_modeling/service_launch_scripts/start_static_retrieval_service.py
@@ -14,8 +14,7 @@
 """
 """
 
-from nemo.collections.nlp.modules.common.megatron.bert_service import start_sentence_bert_server
-from nemo.collections.nlp.modules.common.megatron.retrieval_service import RetrievalServer
+from nemo.collections.nlp.modules.common.megatron.retrieval_services.static_retrieve_server import RetrievalServer
 from nemo.collections.nlp.modules.common.tokenizer_utils import get_nmt_tokenizer
 from nemo.core.config import hydra_runner
